feat: validate some frames & remove WebsocketFrame class
