+++ lib/websocket/connection.js
@@ -3,7 +3,7 @@
 // TODO: crypto isn't available in all environments
 const { randomBytes, createHash } = require('crypto')
 const diagnosticsChannel = require('diagnostics_channel')
-const { uid, states, opcodes } = require('./constants')
+const { uid, states } = require('./constants')
 const {
   kReadyState,
   kResponse,
@@ -11,11 +11,11 @@ const {
   kProtocol,
   kClosingFrame,
   kSentClose,
-  kByteParser
+  kByteParser,
+  kController
 } = require('./symbols')
-const { fireEvent, isEstablished, isClosed, isClosing } = require('./util')
+const { fireEvent, isEstablished } = require('./util')
 const { CloseEvent } = require('./events')
-const { WebsocketFrameSend } = require('./frame')
 const { ByteParser } = require('./receiver')
 const { makeRequest } = require('../fetch/request')
 const { fetching } = require('../fetch/index')
@@ -99,7 +99,7 @@ function establishWebSocketConnection (url, protocols, ws) {
       // 1. If response is a network error or its status is not 101,
       //    fail the WebSocket connection.
       if (response.type === 'error' || response.status !== 101) {
-        failWebsocketConnection(controller)
+        failWebsocketConnection(ws)
         return
       }
 
@@ -108,7 +108,7 @@ function establishWebSocketConnection (url, protocols, ws) {
       //    header list results in null, failure, or the empty byte
       //    sequence, then fail the WebSocket connection.
       if (protocols.length !== 0 && !response.headersList.get('Sec-WebSocket-Protocol')) {
-        failWebsocketConnection(controller)
+        failWebsocketConnection(ws)
         return
       }
 
@@ -123,7 +123,7 @@ function establishWebSocketConnection (url, protocols, ws) {
       //    insensitive match for the value "websocket", the client MUST
       //    _Fail the WebSocket Connection_.
       if (response.headersList.get('Upgrade')?.toLowerCase() !== 'websocket') {
-        failWebsocketConnection(controller)
+        failWebsocketConnection(ws)
         return
       }
 
@@ -132,7 +132,7 @@ function establishWebSocketConnection (url, protocols, ws) {
       //    ASCII case-insensitive match for the value "Upgrade", the client
       //    MUST _Fail the WebSocket Connection_.
       if (response.headersList.get('Connection')?.toLowerCase() !== 'upgrade') {
-        failWebsocketConnection(controller)
+        failWebsocketConnection(ws)
         return
       }
 
@@ -146,7 +146,7 @@ function establishWebSocketConnection (url, protocols, ws) {
       const secWSAccept = response.headersList.get('Sec-WebSocket-Accept')
       const digest = createHash('sha1').update(keyValue + uid).digest('base64')
       if (secWSAccept !== digest) {
-        failWebsocketConnection(controller, response.socket)
+        failWebsocketConnection(ws)
         return
       }
 
@@ -160,7 +160,7 @@ function establishWebSocketConnection (url, protocols, ws) {
       const secExtension = response.headersList.get('Sec-WebSocket-Extensions')
 
       if (secExtension !== null && secExtension !== permessageDeflate) {
-        failWebsocketConnection(controller, response.socket)
+        failWebsocketConnection(ws)
         return
       }
 
@@ -172,7 +172,7 @@ function establishWebSocketConnection (url, protocols, ws) {
       const secProtocol = response.headersList.get('Sec-WebSocket-Protocol')
 
       if (secProtocol !== null && secProtocol !== request.headersList.get('Sec-WebSocket-Protocol')) {
-        failWebsocketConnection(controller, response.socket)
+        failWebsocketConnection(ws)
         return
       }
 
@@ -197,14 +197,15 @@ function establishWebSocketConnection (url, protocols, ws) {
 }
 
 /**
- * @param {import('../fetch/index').Fetch} controller
- * @param {import('net').Socket|import('tls').TLSSocket|undefined} socket
+ * @param {import('./websocket').WebSocket} ws
  */
-function failWebsocketConnection (controller, socket) {
+function failWebsocketConnection (ws) {
+  const { [kController]: controller, [kResponse]: response } = ws
+
   controller.abort()
 
-  if (socket && !socket.destroyed) {
-    socket.destroy()
+  if (!response.socket.destroyed) {
+    response.socket.destroy()
   }
 }
 
@@ -325,51 +326,7 @@ function socketClosed (ws) {
   })
 }
 
-/**
- * @param {WebsocketFrame} frame
- * @param {import('net').Socket} socket
- * @param {import('./websocket').WebSocket} ws
- */
-function handlePing (socket, ws, frame) {
-  //  Upon receipt of a Ping frame, an endpoint MUST send a Pong frame in
-  // response, unless it already received a Close frame.  It SHOULD
-  // respond with Pong frame as soon as is practical.
-
-  if (isClosing(ws) || isClosed(ws)) {
-    return
-  }
-
-  if (channels.ping.hasSubscribers) {
-    channels.ping.publish({ frame })
-  }
-
-  // A Pong frame sent in response to a Ping frame must have identical
-  // "Application data" as found in the message body of the Ping frame
-  // being replied to.
-  const sendFrame = new WebsocketFrameSend(frame.data)
-  const buffer = sendFrame.createFrame(opcodes.PONG)
-
-  socket.write(buffer)
-}
-
-/**
- * @param {WebsocketFrame} frame
- */
-function handlePong (frame) {
-  // A Pong frame sent in response to a Ping frame must have identical
-  // "Application data" as found in the message body of the Ping frame
-  // being replied to.
-
-  if (channels.pong.hasSubscribers) {
-    channels.pong.publish({ frame })
-  }
-
-  // TODO
-}
-
 module.exports = {
   establishWebSocketConnection,
-  failWebsocketConnection,
-  handlePing,
-  handlePong
+  failWebsocketConnection
 }
