+++ lib/websocket/receiver.js
@@ -1,5 +1,8 @@
 const { Writable } = require('stream')
-const { parserStates } = require('./constants')
+const { parserStates, opcodes, states } = require('./constants')
+const { failWebsocketConnection } = require('./connection')
+const { kReadyState } = require('./symbols')
+const { isValidStatusCode } = require('./util')
 
 class ByteParser extends Writable {
   #buffers = []
@@ -43,7 +46,13 @@ class ByteParser extends Writable {
       this.#info.fin = (buffer[0] & 0x80) !== 0
       this.#info.opcode = buffer[0] & 0x0F
 
-      // TODO: HANDLE INVALID OPCODES HERE
+      const fragmented = !this.#info.fin && this.#info.opcode !== opcodes.CONTINUATION
+
+      if (fragmented && this.#info.opcode !== opcodes.BINARY && this.#info.opcode !== opcodes.TEXT) {
+        // Only text and binary frames can be fragmented
+        failWebsocketConnection(this.ws)
+        return
+      }
 
       const payloadLength = buffer[1] & 0x7F
 
@@ -55,6 +64,30 @@ class ByteParser extends Writable {
       } else if (payloadLength === 127) {
         this.#state = parserStates.PAYLOADLENGTH_64
       }
+
+      if (
+        (this.#info.opcode === opcodes.PING ||
+          this.#info.opcode === opcodes.PONG ||
+          this.#info.opcode === opcodes.CLOSE) &&
+        payloadLength > 125
+      ) {
+        // Control frames can have a payload length of 125 bytes MAX
+        failWebsocketConnection(this.ws)
+        return
+      } else if (this.#info.opcode === opcodes.CLOSE) {
+        if (payloadLength === 1) {
+          failWebsocketConnection(this.ws)
+          return
+        }
+
+        // Upon either sending or receiving a Close control frame, it is said
+        // that _The WebSocket Closing Handshake is Started_ and that the
+        // WebSocket connection is in the CLOSING state.
+        this.ws[kReadyState] = states.CLOSED
+      }
+
+      // TODO: handle control frames here. Since they are unfragmented, and can
+      // be sent in the middle of other frames, we shouldn't parse them as normal.
 
       this.#buffers = [buffer.subarray(2)]
       this.#byteOffset -= 2
@@ -102,6 +135,8 @@ class ByteParser extends Writable {
           this.#byteOffset = 0
         }
 
+        // HANDLE FULL FRAME HERE
+
         this.#info = {}
         this.#state = parserStates.INFO
       }
@@ -112,6 +147,51 @@ class ByteParser extends Writable {
     } else {
       callback()
     }
+  }
+
+  parseCloseBody (onlyCode) {
+    // https://datatracker.ietf.org/doc/html/rfc6455#section-7.1.5
+    /** @type {number|undefined} */
+    let code
+
+    const { data } = this.#info
+
+    if (data.length > 2) {
+      // _The WebSocket Connection Close Code_ is
+      // defined as the status code (Section 7.4) contained in the first Close
+      // control frame received by the application
+      code = data.readUInt16BE(0)
+    }
+
+    if (onlyCode) {
+      if (!isValidStatusCode(code)) {
+        return null
+      }
+
+      return { code }
+    }
+
+    // https://datatracker.ietf.org/doc/html/rfc6455#section-7.1.6
+    /** @type {Buffer} */
+    let reason = data.subarray(2)
+
+    // Remove BOM
+    if (reason[0] === 0xEF && reason[1] === 0xBB && reason[2] === 0xBF) {
+      reason = reason.subarray(3)
+    }
+
+    if (code !== undefined && !isValidStatusCode(code)) {
+      return null
+    }
+
+    try {
+      // TODO: optimize this
+      reason = new TextDecoder('utf-8', { fatal: true }).decode(reason)
+    } catch {
+      return null
+    }
+
+    return { code, reason }
   }
 }
