+++ lib/websocket/websocket.js
@@ -176,7 +176,7 @@ class WebSocket extends EventTarget {
       // If the WebSocket connection is not yet established
       // Fail the WebSocket connection and set this's ready state
       // to CLOSING (2).
-      failWebsocketConnection(this[kController])
+      failWebsocketConnection(this)
       this[kReadyState] = WebSocket.CLOSING
     } else if (!isClosing(this)) {
       // If the WebSocket closing handshake has not yet been started
